Move some enum and literal related type ops to typeops.py
I keep wanting access to these functions outside of checker.py
in various diffs I'm working on.

So as suggested earlier, I might as well just do this refactoring
now, independent of the other PRs.
